backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "InterXAI - AI-powered interview automation system (FastAPI backend)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0",
    "fastapi>=0.115.0",
    "greenlet>=3.4.0",
    "pyjwt>=2.9.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "ruff>=0.15.10",
]
